// src/api/nexuscore/model/rulesColumnsBlock.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Nexuscore
 * The backend for all Everthorn internal services
 * OpenAPI spec version: 1.0.0
 */

export interface RulesColumnsBlock {
  /** The block id */
  block_id: number;
  /** The fractional ordering index of this block on the page */
  position: number;
  /** ISO 8601 datetime the block was created */
  created_at: string;
  /** ISO 8601 datetime the block was last updated */
  updated_at: string;
  /** The type of the block */
  type: 'rules_columns';
  /** Allowed actions/behaviours during the event */
  allowed: string[];
  /** Disallowed actions/behaviours during the event */
  disallowed: string[];
}